refactor: extract NavigationThemeProvider and apply to tabs layout
