+++ app/(tabs)/_layout.tsx
@@ -1,3 +1,4 @@
+import { NavigationThemeProvider } from '@/components/navigation/NavigationThemeProvider';
 import NativeTabsLayout from '@/components/tabs/NativeTabsLayout';
 import OldLegacyTabs from '@/components/tabs/OldLegacyTabs';
 import { Platform } from 'react-native';
@@ -7,5 +8,9 @@ export default function TabLayout() {
     return <OldLegacyTabs />;
   }
 
-  return <NativeTabsLayout />;
+  return (
+    <NavigationThemeProvider>
+      <NativeTabsLayout />
+    </NavigationThemeProvider>
+  );
 }
